Bigger Changes - Parseing is moved into the controller - All shared elements are now in the controller decorator - Save and load are moved into the main controller - Delete class now also removes all relations that class was a part of
