+++ src/Models/__init__.py
@@ -9,6 +9,5 @@ from .uml_param import *
 from .uml_relation import *
 from .uml_visitor import *
 from .uml_save_load import *
-from .uml_parser import parse
 from .error_handler import Error_Handler
 from .uml_undo_redo import UML_States
